Update unstructured/native grid plotting
